deploy/rulegen/audit_logging_test.go
package rulegen

import (
	"testing"

	"github.com/google/go-cmp/cmp"
	"gopkg.in/yaml.v2"
)

func TestAudLoggingRules(t *testing.T) {
	config, _ := getTestConfigAndProject(t, nil)
	got, err := AuditLoggingRules(config)
	if err != nil {
		t.Fatalf("AuditLoggingRules = %v", err)
	}

	wantYAML := `
- name: Require all Cloud Audit logs.
  resource:
  - type: project
    resource_ids:
    - '*'
  service: allServices
  log_types:
  - ADMIN_READ
  - DATA_READ
  - DATA_WRITE
`
	want := make([]AuditLoggingRule, 1)
	if err := yaml.Unmarshal([]byte(wantYAML), &want); err != nil {
		t.Fatalf("yaml.Unmarshal = %v", err)
	}

	if diff := cmp.Diff(got, want); diff != "" {
		t.Errorf("rules differ (-got, +want):\n%v", diff)
	}
}
